reuploaded hompeage
reuploaded homepage because there was a problem when I tried to pull to master. I think it is because I worked directly off of travel-fly hero and copied the content to a new folder for the assignment. I renamed the travel fly hero folder to travel-fly homepage so hopefully this works